refactor: remove legacy anthropic helper entrypoint
